+++ R/gdc-package.R
@@ -16,6 +16,8 @@
 #' @section downloading data:
 #' data
 #'
+#' @importFrom magrittr "%>%"
+#'
 #' @docType package
 #' @name GenomicDataCommons
 NULL