+++ R/simplify_results.R
@@ -49,7 +49,7 @@ rbindlist2 = function(x,...) {
 #' @return a data.frame, potentially with list columns still
 #'     present.
 #'
-#' @examples 
+#' @examples
 #' expands = c("diagnoses","diagnoses.treatments","annotations",
 #'             "demographic","exposures")
 #' head(cases() %>% expand(expands) %>% results() %>% as.data.frame())