+++ NEWS.md
@@ -1,4 +1,9 @@
-# GenomicDataCommons 0.99.8
+# GenomicDataCommons 0.99.12
+
+* Added `readHTSeqFile` for rapid reading of HTSeq-counts output files
+
+
+# GenomicDataCommons 0.99.9
 
 ## API changes
 
@@ -10,11 +15,20 @@
 
 * `as.data.frame()` allows a robust if not entirely complete conversion
   of `GDCResults` to `data.frame`.
+* Added first of data helper functions, `readDNAcopy`, for dealing
+  with files returned from GDC.
 
+## Other
+
+* `options('gdc.verbose'=TRUE)` now dumps URL for gdcdata
+* `ids.GDCQuery` now returns all entity ids and not just the first
+  page of results.
+* test fix for change in return value of `status()` endpoint
 
 ## Bug Fixes
 
 * many--this is beta software
+
 
 # GenomicDataCommons 0.99.6